### 2021-09-03 #3681 @whedon
[PRE REVIEW]: AMAT: A Python package for rapid conceptual design of aerocapture and atmospheric Entry, Descent, and Landing (EDL) missions in a Jupyter environment
**Submitting author:** @athulpg007 (<a href="http://orcid.org/0000-0002-9326-3293">Athul Pradeepkumar Girija</a>)
**Repository:** <a href="https://github.com/athulpg007/AMAT" target ="_blank">https://github.com/athulpg007/AMAT</a>
**Version:** v2.1.8
