### 2021-09-03 #3681 @kyleniemeyer
Hi @jbytecode, could you edit this one? I realize the application may be a bit outside your wheelhouse, but the statistical side of things seems right. Thanks
