### 2021-09-04 7630e1d
Add missing DOI to paper.bib
