+++ paper/paper.bib
@@ -54,7 +54,8 @@ title          = "{Vision and Voyages for Planetary Science in the Decade 2013-2
 address        = "{Washington D.C.}",
 chapter        = "{5}",
 year           = 2011,
-url =          {https://www.nap.edu/catalog/13117/vision-and-voyages-for-planetary-science-in-the-decade-2013-2022}
+url =          {https://www.nap.edu/catalog/13117/vision-and-voyages-for-planetary-science-in-the-decade-2013-2022},
+doi            = "10.17226/13117"
 }
 
 @TECHREPORT{Leszczynski1998,
