### 2021-09-05 a87b15c
Fix typo in paper.md
