### 2021-09-09 #3681 @jbytecode
👋👋👋 @helgee  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "_AMAT: A Python package for rapid conceptual design of aerocapture and atmospheric Entry, Descent, and Landing (EDL) missions in a Jupyter environment_". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3681).
