### 2021-09-09 #3681 @RobertRosca
Hey! Oooh, this seems super interesting, I'm up for reviewing it but I already have https://github.com/openjournals/joss-reviews/issues/3641#issuecomment-914074505 waiting and haven't had a chance to look at it yet since I've been pretty busy.

I'll do my best to finish #3641 asap and if I manage to do it before somebody is found for this, then I'll accept.
