### 2021-09-10 athulpg007/AMAT#3 @astrojuanlu
Consider relicensing
> We recommend against using Creative Commons licenses for software. Instead, we strongly encourage you to use one of the very good software licenses which are already available. We recommend considering licenses listed as free by the Free Software Foundation and listed as “open source” by the Open Source Initiative.

https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software

More information: https://opensource.stackexchange.com/a/1718

Given that you are using CC-BY-SA, I would recommend a copyleft license, like GPLv3 (or "GPLv3 or later"), see https://choosealicense.com/licenses/gpl-3.0/
