### 2021-09-10 athulpg007/AMAT#4 @astrojuanlu
Add API documentation?
The documentation contains a lot of comprehensive examples, that's awesome! 🚀

I wonder if you would consider adding an API reference as well, using [autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html), [autosummary](https://www.sphinx-doc.org/en/master/usage/extensions/autosummary.html), or [sphinx-autoapi](https://pypi.org/project/sphinx-autoapi/)?
