### 2021-09-10 #3710 @astrojuanlu
I did a first pass, and opened https://github.com/athulpg007/AMAT/issues/3 and https://github.com/athulpg007/AMAT/issues/4 accordingly. In addition, I didn't see automated tests or contribution instructions, so for now I left those checkboxes unchecked. I still have to go over the rest.
