### 2021-09-11 athulpg007/AMAT#5 @helgee
Clean-up dependencies
CC: @astrojuanlu

Since `pandas` and `jupyterlab` are only needed for running the examples and not called from inside the AMAT package, they should not be hard dependencies. Especially installing JupyterLab is a bit excessive.

I would recommend to remove them from [setup.py](https://github.com/athulpg007/AMAT/blob/master/setup.py#L16) and rather instruct users to install them manually if they want to run example notebooks.

I also noticed that `setup.py` and `requirements.txt` are out-of-sync. But I must confess that my Python dependencies management knowledge is a bit rusty and I am not sure what the best practice here is.
