### 2021-09-11 1dc2979
Added API reference section to docs, clean build
