+++ docs/build/html/_sources/jupyter-notebooks.rst.txt
@@ -1,4 +1,0 @@
-.. toctree::
-   :maxdepth: 1
-
-   _examples/example-01-hello-world.ipynb
