+++ docs/build/html/_sources/jupyter.rst.txt
@@ -1,4 +1,0 @@
-.. toctree::
-   :maxdepth: 1
-
-   /examples/example-01-hello-world.ipynb
