+++ docs/source/api_reference.rst
@@ -1,0 +1,12 @@
+API Reference
+=============
+
+.. toctree::
+   :maxdepth: 4
+
+.. automodule:: AMAT.planet
+   :members:
+.. automodule:: AMAT.vehicle
+   :members:
+.. automodule:: AMAT.launcher
+   :members:
