### 2021-09-11 54c7b69
Remove pandas and jupyerlab hard dependency requirement, update to v2.1.9
