### 2021-09-11 b649b31
Update docs to instruct users to install jupyterlab and pandas to run examples
