+++ docs/source/AMAT.rst
@@ -28,7 +28,6 @@ AMAT.vehicle module
    :undoc-members:
    :show-inheritance:
 
-
 Module contents
 ---------------
 
