+++ setup.py
@@ -11,7 +11,7 @@ setup(name='AMAT',
       long_description_content_type="text/markdown",
       author='Athul P. Girija',
       author_email='athulpg007@gmail.com',
-      license='cc-by-sa-4.0',
+      license='GPL-3.0-or-later',
       packages=['AMAT'],
       install_requires=['numpy', 'scipy', 'matplotlib'],
       classifiers=[
