### 2021-09-11 athulpg007/AMAT#3 @athulpg007
Changed license to GPL-3.0-or-later as suggested.
