### 2021-09-11 athulpg007/AMAT#4 @athulpg007
Added an [API Reference](https://amat.readthedocs.io/en/master/api_reference.html) section to docs.  I was not entirely sure to make it more concise looking. I can make further changes if needed.
