### 2021-09-11 athulpg007/AMAT#5 @athulpg007
Removed pandas and jupyterlab from setup.py. Added a note in installation instructions telling users to install them manually to run the example notebooks.

The requirements.txt was made for readthedocs build. I am not sure of what the best practice is. I know the rtd build works only with some frozen package versions in the requirements.txt.
