### 2021-09-11 #3710 @athulpg007
I have addressed [athulpg007/AMAT/#/3](https://github.com/athulpg007/AMAT/#/3) and [athulpg007/AMAT/#/4](https://github.com/athulpg007/AMAT/#/4).

I will add some tests and contribution instructions later today. Thanks!
