### 2021-09-11 athulpg007/AMAT#5 @helgee
> The requirements.txt was made for readthedocs build. I am not sure of what the best practice is. I know the rtd build works only with some frozen package versions in the requirements.txt.

@astrojuanlu Can you comment on this?
