### 2021-09-11 athulpg007/AMAT#5 @athulpg007
While I wait for the reply, I will look at some other repositories and the readthedocs docs on how to do this in a cleaner way.
