+++ tests/test_planet.py
@@ -1,0 +1,222 @@
+"""
+test_planet.py
+
+Tests to check if Planet objects can be created and loaded with
+atmospheric profiles.
+
+"""
+
+import unittest
+
+try:
+    from AMAT.planet import Planet
+except ModuleNotFoundError:
+    raise ModuleNotFoundError("Cannot import Planet from AMAT.planet")
+
+class TestPlanet(unittest.TestCase):
+
+	def test_create_planet(self):
+		"""
+		Test that if a Planet object instances can be created for
+		various Solar System destinations
+		"""
+		planet1 = Planet("VENUS")
+		message = "Could not create Planet instance for Venus"
+		self.assertIsInstance(planet1, Planet, message)
+
+		planet2 = Planet("EARTH")
+		message = "Could not create Planet instance for Earth"
+		self.assertIsInstance(planet2, Planet, message)
+
+		planet3 = Planet("MARS")
+		message = "Could not create Planet instance for Mars"
+		self.assertIsInstance(planet3, Planet, message)
+
+		planet4 = Planet("JUPITER")
+		message = "Could not create Planet instance for Jupiter"
+		self.assertIsInstance(planet4, Planet, message)
+
+		planet5 = Planet("SATURN")
+		message = "Could not create Planet instance for Saturn"
+		self.assertIsInstance(planet5, Planet, message)
+
+		planet6 = Planet("TITAN")
+		message = "Could not create Planet instance for Saturn"
+		self.assertIsInstance(planet6, Planet, message)
+
+		planet7 = Planet("URANUS")
+		message = "Could not create Planet instance for URANUS"
+		self.assertIsInstance(planet7, Planet, message)
+		
+		planet8 = Planet("NEPTUNE")
+		message = "Could not create Planet instance for Saturn"
+		self.assertIsInstance(planet8, Planet, message)
+
+		"""
+		check GM values for all planet objects
+		"""
+
+		self.assertEqual(planet1.GM, 3.248599E14, msg="Check Venus GM")
+		self.assertEqual(planet2.GM, 3.986004E14, msg="Check Earth GM")
+		self.assertEqual(planet3.GM, 4.282837E13, msg="Check Mars GM")
+		self.assertEqual(planet4.GM, 1.26686534E17, msg="Check Jupiter GM")
+		self.assertEqual(planet5.GM, 3.7931187E16, msg="Check Saturn GM")
+		self.assertEqual(planet6.GM, 8.9780000E12, msg="Check Titan GM")
+		self.assertEqual(planet7.GM, 5.793939E15, msg="Check Uranus GM")
+		self.assertEqual(planet8.GM, 6.8365299E15, msg="Check Neptune GM")
+
+		"""
+		check h_thres is set to a sufficiently high value for all planets
+		"""
+		self.assertGreaterEqual(planet1.h_thres, 100.0E3, msg="Venus atm. cut off height likely too low.")
+		self.assertGreaterEqual(planet2.h_thres, 100.0E3, msg="Earth atm. cut off height likely too low.")
+		self.assertGreaterEqual(planet3.h_thres, 100.0E3, msg="Mars atm. cut off height likely too low.")
+		self.assertGreaterEqual(planet4.h_thres, 800.0E3, msg="Jupiter atm. cut off height likely too low.")
+		self.assertGreaterEqual(planet5.h_thres, 800.0E3, msg="Saturn atm. cut off height likely too low.")
+		self.assertGreaterEqual(planet6.h_thres, 600.0E3, msg="Titan atm. cut off height likely too low.")
[…]
