+++ tests/__init__.py
@@ -1,0 +1,3 @@
+from tests import test_planet
+from tests import test_vehicle
+from tests import test_launcher
