+++ MANIFEST.in
@@ -4,6 +4,7 @@ include LICENSE
 # Include the data files
 recursive-include examples *
 recursive-include plots *
+recursive-include tests *
 
 # Exclude atm-data files
 recursive-exclude atmdata *

[…]
