### 2021-09-19 4d8c7ca
Add contribution instructions to README.md and docs/
