+++ docs/source/contributions.rst
@@ -1,0 +1,17 @@
+Contributions
+=============
+
+If you wish to contribute or report an issue, feel free to e-mail me here_ or to open an issue_ or create a pull_ request here from the repository_.
+
+If you wish to make a contribution, you can do as follows:
+
+ * fork the GitHub repository 
+ * create a feature branch from *master* 
+ * add your feature and document it
+ * run the tests 
+ * open a pull request
+
+.. _here: mailto:athulpg007@gmail.com
+.. _issue: https://github.com/athulpg007/AMAT/issues
+.. _pull: https://github.com/athulpg007/AMAT/pulls
+.. _repository: https://github.com/athulpg007/AMAT
