### 2021-09-19 #3710 @jbytecode
@athulpg007 thank you for adding tests. It is also a good practice to add a GitHub action that performs the tests after a commit so people can see the latest changes don't break the package.
