### 2021-09-25 3de74c5
Update README.md and other files for GNU GPLv3 license
