### 2021-09-25 7d0cf07
Update .readthedocs.yml to V2
