+++ .readthedocs.yml
@@ -1,3 +1,14 @@
+version: 2
+
+# Build documentation in the docs/ directory with Sphinx
+sphinx:
+   configuration: docs/conf.py
+
+# Optionally build your docs in additional formats such as PDF
+formats:
+   - pdf
+
+
 python:
-  version: 3.7
-  setup_py_install: true
+  install:
+    - method: pip
