### 2021-09-25 35c0008
Add path variable to .readthedocs.yml
