+++ .readthedocs.yml
@@ -12,3 +12,4 @@ formats:
 python:
   install:
     - method: pip
+    path: .
