### 2021-09-25 e43a9cb
Moved requirements file to docs/
