+++ docs/requirements.txt
@@ -1,3 +1,5 @@
+# Documentation Dependencies
+
 sphinx==2.1.2
 ipykernel==5.1.1
 nbsphinx==0.6.1
