### 2021-09-25 4b4b30c
Update dependencies in docs/requirements.txt
