+++ .readthedocs.yml
@@ -10,3 +10,4 @@ formats: all
 python:
   install:
     - requirements: docs/requirements.txt
+    - method: pip
