+++ .readthedocs.yml
@@ -9,5 +9,7 @@ formats: all
 
 python:
   install:
-    - requirements: docs/requirements.txt
-    - method: pip
+      - method: pip
+      - path: .
+      - requirements: docs/requirements.txt
+
