+++ .readthedocs.yml
@@ -10,6 +10,6 @@ formats: all
 python:
   install:
       - method: pip
-      - path: .
+        path: .
       - requirements: docs/requirements.txt
 
