+++ .readthedocs.yml
@@ -2,7 +2,7 @@ version: 2
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   configuration: docs/conf.py
+   configuration: docs/source/conf.py
 
 # Optionally build your docs in additional formats such as PDF
 formats: all
