### 2021-09-25 9e45156
Remove jinja2 pinned version in docs/requirements.txt
