+++ docs/requirements.txt
@@ -8,5 +8,5 @@ matplotlib
 
 sphinx==2.1.2
 ipykernel==5.1.1
-nbsphinx==0.6.1
-#jinja2==2.11.3
+nbsphinx==0.8.6
+jinja2==3.0.1
