### 2021-09-25 b52f21e
Add  jinja2 pinned version in docs/requirements.txt
