### 2021-09-25 a4941e4
Revert docs/requirements.txt to working pinned dep. versions
