### 2021-09-25 4ce0d3e
Pin dependency versions in docs/requirements.txt
