### 2021-09-25 e304869
Pin sphinx==4.1.2 in docs/requirements.txt
