### 2021-09-25 8e594a4
Try pinning versions in docs/requirements.txt for rtd build
