+++ docs/requirements.txt
@@ -7,6 +7,6 @@ matplotlib
 # Documentation Dependencies
 
 sphinx==4.1.2
-ipykernel==5.1.1
+ipykernel==6.4.1
 nbsphinx==0.6.1
-jinja2==2.11.3
+jinja2==3.0.1
