### 2021-09-25 48ba46b
Try pinning versions in docs/requirements.txt for rtd build
