+++ docs/requirements.txt
@@ -8,5 +8,5 @@ matplotlib
 
 sphinx==4.1.2
 ipykernel==6.4.1
-nbsphinx==0.6.1
+nbsphinx==0.8.6
 jinja2==3.0.1
