### 2021-09-25 cb9f5ca
Revert to pinned versions in docs/requirements.txt for rtd build
