### 2021-09-26 dd7c095
Create GitHub action for testing
Add .github/workflows/pytest.yml
