+++ .github/workflows/pytest.yml
@@ -33,4 +33,5 @@ jobs:
       #  flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
+        cd $GITHUB_WORKSPACE/tests/
         pytest
