### 2021-09-26 athulpg007/AMAT#5 @athulpg007
Moved the requirements file to docs/requirements.txt with dependencies for both the package and docs now listed in a single file.
