### 2021-09-26 #3710 @athulpg007
License has been updated to GPL-v3-or-later. Updated up header text from some files with the old CC-BY-SA-4.0 license text to reflect the new license. If this issue if resolved, I will close [athulpg007/AMAT/#3](https://github.com/athulpg007/AMAT/issues/3).

Removed pandas and jupyterlab from setup.py Listed package and readthedocs build dependencies in docs/requirements.txt. See if this resolves [athulpg007/AMAT/#5](https://github.com/athulpg007/AMAT/issues/5)

Added GitHub action to perform tests after a commit.
