### 2021-09-26 #3710 @jbytecode
> Added GitHub action to perform tests after a commit.

Great! Thank you. @athulpg007
