+++ docs/requirements.txt
@@ -9,4 +9,4 @@ matplotlib
 sphinx==2.1.2
 ipykernel==5.1.1
 nbsphinx==0.6.1
-jinja2==3.0.1
+jinja2==2.11.3
