### 2021-10-07 #3710 @helgee
Quick note @athulpg007: I have verified that the Linux installation instructions are valid for macOS as well.
