### 2021-10-07 437dc23
Add Linux/MacOS to installation instructions
