### 2021-11-01 athulpg007/AMAT#5 @astrojuanlu
Having a `docs/requirements.txt` with the documentation-related dependencies is [the recommended solution](https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html). This can be closed 👍🏽
