### 2021-11-02 284b895
JOSS editor updates on paper.md
