### 2021-11-02 c2f91e5
JOSS Editor Update for paper.md (2)
