### 2021-11-04 1870bd3
Update references in paper.md
