+++ paper/paper.bib
@@ -54,8 +54,7 @@ title          = "{Vision and Voyages for Planetary Science in the Decade 2013-2
 address        = "{Washington D.C.}",
 chapter        = "{5}",
 year           = 2011,
-url =          {https://www.nap.edu/catalog/13117/vision-and-voyages-for-planetary-science-in-the-decade-2013-2022},
-doi            = "10.17226/13117"
+url =          {https://www.nap.edu/catalog/13117/vision-and-voyages-for-planetary-science-in-the-decade-2013-2022}
 }
 
 @TECHREPORT{Leszczynski1998,
@@ -77,7 +76,7 @@ year          = "1998"
 }
 
 @techreport{Hofstadter2017,
-author       = "{Hofstadter}, Mark D and Simon, Amy and Reh, Kim and Elliot, John and {et al.}",
+author       = "{Hofstadter}, Mark D and Simon, Amy and Reh, Kim and {et al.}",
 institution  = {NASA},
 pages        = {473--488},
 title        = "{Ice Giants Pre-Decadal Study Final Report}",
@@ -177,7 +176,7 @@ doi       = "10.2514/1.A33997"
 
 @MISC{hash2007fire,
   title   =  "{FIRE II Calculations for Hypersonic Nonequilibrium Aerothermodynamics Code Verification: DPLR, LAURA, and US3D}",
-  author  = "{Hash}, David and {Olejniczak}, Joseph and Wright, Michael and Prabhu, Dinesh and Pulsonetti, Maria and Hollis, Brian and Gnoffo, Peter and Barnhardt, Michael and Nompelis, Ioannis and Candler, Graham",
+  author  = "{Hash}, David and {Olejniczak}, Joseph and Wright, Michael and Prabhu, Dinesh and {et al.}",
   booktitle  ={45th AIAA Aerospace Sciences Meeting and Exhibit},
   pages      ={605},
   year       ={2007},
@@ -198,7 +197,7 @@ booktitle = "{Planetary Probe Atmospheric Entry and Descent Trajectory Analysis 
 
 @article{Saikia2021,
 title = "{Aerocapture Assessment for NASA Ice Giants Pre-Decadal Survey Mission Study}",
-author ="{Saikia}, Sarag J and Millane, James and Lu, Ye and Mudek, Alec and Arora, Archit and Witsberge, Paul and Hughes, Kyle and Longuski, James M and Spilker, Thomas and Petropoulos, Anastassios and Arora, Nitin and Cutts, James  and Elliott, John and Sims, Jon and Reh, Kim",
+author ="{Saikia}, Sarag J and Millane, James and Lu, Ye and Mudek, Alec and Arora, Archit and {et al.}",
 journal= "{Journal of Spacecraft and Rockets}",
 volume = 58,
 number = 2,
