### 2021-11-04 fdbecbd
Update PyPI and docs version to v2.2.1
