+++ build/lib/AMAT/planet.py
@@ -1,43 +1,21 @@
-# Author(s) : Athul Pradeepkumar Girija; apradee@purdue.edu
-#
-# License : CC-BY-SA-4.0
-#
-# You are free to:
-#
-# Share — copy and redistribute the material in any medium or format
-# Adapt — remix, transform, and build upon the material for any#
-# purpose,  even commercially.
+"""
+GNU GENERAL PUBLIC LICENSE
+Version 3, 29 June 2007
 
-# The licensor cannot revoke these freedoms as long as you follow the 
-# license terms
+AMAT provides mission analysis capability for aerocapture and Entry, 
+Descent, and Landing (EDL) concepts for planetary science missions.
 
-# Attribution — You must give appropriate credit, provide a link to 
-# the license, and indicate if changes were made. You may do so in any
-# reasonable manner, but not in any way that suggests the licensor 
-# endorses you or your use.
+Copyright (C) 2021  Athul Pradeepkumar Girija.
 
-# ShareAlike — If you remix, transform, or build upon the material, 
-# you must distribute your contributions under the same license as 
-# the original.  
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License Version 3 as 
+published by the Free Software Foundation.
 
-# The above copyright notice and this permission notice shall be 
-# included in all copies or substantial portions of the Software.
-
-# ===================================================================
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, 
-# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF 
-# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND 
-# NONINFRINGEMENT. 
-
-# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY 
-# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, 
-# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE 
-# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-# ===================================================================
-
-# This script requires that `numpy`, 'scipy', and ''matplotlib
-# be installed within  the Python environment you are running 
-# this script in.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+"""
 
 import numpy as np
 from scipy.interpolate import interp1d
