### 2021-11-04 #3710 @jbytecode
@athulpg007

Could you please correct me if I am wrong with the line 60:

-  Figure 3 shows example trajectories for a drag modulation system at Mars computed by using AMAT.

means

-  Figure 3 shows `an` example `of` trajectories for a drag modulation system at Mars computed by using AMAT.

or something else? If yes, please correct all of the Figure descriptions. Thank you.
