### 2021-11-05 b3c3c11
Corrections to paper.md
