+++ paper/paper.md
@@ -35,30 +35,30 @@ Software tools for identification of trajectories and techniques that enhance or
 
 # Rapid Design Capability
 
-AMAT can be used to compute aerocapture trajectories as well as key design parameters such as theoretical corridor width (TCW) and stagnation-point heat rate using just a few lines of code. Figure 1 shows the undershoot and overshoot trajectories for lift modulation aerocapture at Venus, and the stagnation-point heat rates computed by using AMAT.
+AMAT can be used to compute key aerocapture design parameters such as theoretical corridor width (TCW) and stagnation-point heat rate using just a few lines of code. Figure 1 shows an example of undershoot and overshoot trajectories for lift modulation aerocapture at Venus, and the corresponding stagnation-point heat rate computed by using AMAT.
 
 ![](https://i.imgur.com/3XPh6JY.png)
-**Figure 1.** Venus aerocapture vehicle trajectory and stagnation-point heat rate computed using AMAT. Figure based on the parameters listed in [@Craig2005].
+**Figure 1.** An example of a Venus aerocapture vehicle trajectory and the stagnation-point heating profile computed by using AMAT. Figure based on [@Craig2005].
 
 AMAT can be used to create **aerocapture feasibility charts**, a graphical method to visualize the mission design trade space considering both interplanetary and vehicle design aspects [@Lu2018Titan]. Aerocapture feasibility charts help the mission designer assess the feasible range of vehicle lift-to-drag ratio ($L/D$) or ballistic-coefficient ratio ($\beta_2/\beta_1$) for lift and drag modulation respectively, as well as the feasible range of interplanetary arrival $V_{\infty}$ considering corridor width, deceleration, peak-heat rate, and total heat load. Figure 2 shows an example feasibility chart for lift modulation aerocapture at Neptune created by using AMAT.
 
 ![](https://i.imgur.com/BNINxh4.png)
 **Figure 2.** Interplanetary trajectory trade space (left, generated using STOUR code) and aerocapture feasibility chart for Neptune (right, generated using AMAT). Figure based on the mission design presented in [@Saikia2021; @Girija2020b].  
 
-AMAT can also be used to quickly set up and compute single-event jettison drag modulation aerocapture corridor bounds, and propagate guided trajectories. Figure 3 shows example trajectories for a drag modulation system at Mars computed by using AMAT.
+AMAT can also be used to quickly set up and compute single-event jettison drag modulation aerocapture corridor bounds, and propagate guided trajectories. Figure 3 shows an example of trajectories for a drag modulation system at Mars computed by using AMAT.
 
 ![](https://i.imgur.com/YlMk6Th.png)
 **Figure 3.** Altitude-velocity profiles and deceleration profiles for a drag modulation aerocapture system at Mars. Solid and dashed lines correspond to a 450 km circular and 1-sol target orbit respectively. Figure based on the flight system presented in [@Werner2019].  
 
-AMAT can be used to perform Monte Carlo simulations to assess user-defined guidance schemes and flight system performance with a baseline vehicle design and interplanetary trajectory considering navigation, atmospheric, and aerodynamic uncertainties.
+AMAT can be used to perform Monte Carlo simulations to assess user-defined guidance schemes and system performance with a vehicle design and interplanetary trajectory considering navigation, atmospheric, and aerodynamic uncertainties as shown in Figure 4.
 
 ![](https://i.imgur.com/Jefki5T.png)
-**Figure 4.** Monte Carlo simulation results showing post-aerocapture orbit dispersion (before the periapsis-raise maneuever) for a drag modulation system at Venus (left) and a lift modulation system at Neptune (right). Based on  [@Girija2020b].
[…]
