### 2021-11-06 f8e346e
Delete unused data directories
