+++ data/girijaSaikia2020a/MCB1/#53/efpa.txt
@@ -1,1 +1,0 @@
--1.383959979237502402e+01
