+++ data/girijaSaikia2020a/MCB1/#52/efpa.txt
@@ -1,1 +1,0 @@
--1.379490008279297264e+01
