+++ data/girijaSaikia2020a/MCB1/#29/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.817800729064614279e+05
