+++ data/girijaSaikia2020a/MCB1/#66/efpa.txt
@@ -1,1 +1,0 @@
--1.385933605160012050e+01
