+++ data/girijaSaikia2020a/MCB1/#61/terminal_periapsis.txt
@@ -1,1 +1,0 @@
-2.193582987939268492e+02
