+++ data/girijaSaikia2020a/MCB1/#21/atmSigma.txt
@@ -1,1 +1,0 @@
--4.224066575461959266e-02
