+++ data/girijaSaikia2020a/MCB1/#37/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.642388750495546847e+05
