+++ data/girijaSaikia2020a/MCB1/#14/periapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-9.664143560169077318e+01
