+++ data/girijaSaikia2020a/MCB1/#31/efpa.txt
@@ -1,1 +1,0 @@
--1.408704899899666252e+01
