+++ data/girijaSaikia2020a/MCB1/#87/apoapsis_error.txt
@@ -1,1 +1,0 @@
--3.005068751991915477e+00
