+++ data/girijaSaikia2020a/MCB1/#76/apoapsis_error.txt
@@ -1,1 +1,0 @@
--6.571988818355239736e+00
