+++ data/girijaSaikia2020a/MCB1/#90/periapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-9.839435027853323845e+01
