### 2021-11-06 680a487
Delete unused data directories
