+++ data/girijaSaikia2020b/MCB1/#66/terminal_periapsis.txt
@@ -1,1 +1,0 @@
-7.907636155497655750e+01
