+++ data/girijaSaikia2020b/MCB1/#55/atmSigma.txt
@@ -1,1 +1,0 @@
-1.367090364811126335e+00
