+++ data/girijaSaikia2020b/MCB1/#75/apoapsis_error.txt
@@ -1,1 +1,0 @@
--5.774048484130718251e+00
