+++ data/girijaSaikia2020b/MCB1/#43/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.623481011215204489e+05
