+++ data/girijaSaikia2020b/MCB1/#53/apoapsis_error.txt
@@ -1,1 +1,0 @@
--8.237904722710183592e+00
