+++ data/girijaSaikia2020b/MCB1/#86/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.909870015027269837e+05
