+++ data/girijaSaikia2020b/MCB1/#10/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.401405967443314148e+05
