+++ data/girijaSaikia2020b/MCB1/#90/LD.txt
@@ -1,1 +1,0 @@
-3.808733516139363173e-01
