+++ data/girijaSaikia2020b/MCB1/#84/apoapsis_raise_DV.txt
@@ -1,1 +1,0 @@
--4.395231661150319269e+00
