+++ data/girijaSaikia2020b/MCB1/#57/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-3.721668893255289877e+05
