+++ data/girijaSaikia2020b/MCB1/#92/LD.txt
@@ -1,1 +1,0 @@
-3.828590862056376332e-01
