+++ data/girijaSaikia2020b/MCB1/#43/apoapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-6.469485401179554174e+01
