### 2021-11-06 cc0ef0d
Delete unused data directories
