+++ data/werner2019/MCB1-Mars/#111/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-4.380427285174694703e+02
