+++ data/werner2019/MCB1/#29/periapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-3.109416411948222958e+01
