+++ data/werner2019/MCB1/#147/efpa.txt
@@ -1,1 +1,0 @@
--4.899044463805024563e+00
