+++ data/werner2019/MCB1/#149/atmSigma.txt
@@ -1,1 +1,0 @@
-1.767590855315989140e+00
