+++ data/werner2019/MCB1/#81/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-1.780047894143301164e+03
