+++ data/werner2019/MCB1/#143/terminal_apoapsis.txt
@@ -1,1 +1,0 @@
-1.298197626038929457e+03
