+++ data/werner2019/MCB2-Mars/#129/efpa.txt
@@ -1,1 +1,0 @@
--1.167972023576653484e+01
