+++ data/werner2019/MCB1-Mars/#104/periapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-1.032053233938213452e+02
