+++ data/werner2019/MCB1-Mars/#4/periapsis_raise_DV.txt
@@ -1,1 +1,0 @@
-1.173007887527201092e+02
