+++ data/werner2019/MCB2-Mars/#69/terminal_periapsis.txt
@@ -1,1 +1,0 @@
-4.532133614064706251e+01
