+++ data/werner2019/MCB1/#170/efpa.txt
@@ -1,1 +1,0 @@
--4.837084447290659561e+00
