### 2021-11-06 41389bc
Corrections to paper.md
