### 2021-11-06 a07079a
JOSS Correction - add missing doi and comma
