+++ paper/paper.bib
@@ -54,7 +54,7 @@ title          = "{Vision and Voyages for Planetary Science in the Decade 2013-2
 address        = "{Washington D.C.}",
 chapter        = "{5}",
 year           = 2011,
-doi            = "10.17226/13117"
+doi            = "10.17226/13117",
 url =          {https://www.nap.edu/catalog/13117/vision-and-voyages-for-planetary-science-in-the-decade-2013-2022}
 }
 
